requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
MouseInfo==0.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pynput==1.8.1
pyobjc-core==11.1
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreText==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
rubicon-objc==0.5.1
scipy==1.13.1
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tinycss2==1.4.0
tk==0.1.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.23.0
